Add tools to validate the derivatives produced by JuMP
