+++ src/SymbolicAD.jl
@@ -2,13 +2,15 @@ module SymbolicAD
 
 import Base.Meta: isexpr
 import JuMP
+import LinearAlgebra
 import MathOptInterface
 import SparseArrays
 import Symbolics
+import Test
 
 const MOI = MathOptInterface
 
 include("nonlinear_oracle.jl")
 include("MOI_wrapper.jl")
-
+include("jump_test_utilities.jl")
 end
